v.0.0.2 - Alpha
This contains a background image, but not all of the code just yet, some issues may still be present, such as your stick figure not showing up. This also contains door image files for later, the door has not been added in yet, I do not know if the stick figure issue is because I made a mistake or if I did not finish coding it.